move array evaluation to array mod
