+++ src/executor_v2/mod.rs
@@ -26,7 +26,6 @@ use self::create::*;
 use self::delete::*;
 use self::drop::*;
 use self::evaluator::*;
-// use self::dummy_scan::*;
 use self::explain::*;
 use self::filter::*;
 use self::hash_agg::*;
@@ -58,8 +57,7 @@ mod copy_to_file;
 mod create;
 mod delete;
 mod drop;
-// mod dummy_scan;
-mod evaluator;
+pub mod evaluator;
 mod explain;
 mod filter;
 mod hash_agg;
